add comments, minor refactoring
